+++ rtengine/capturesharpening.cc
@@ -583,10 +583,10 @@ BENCHFUN
                 } else {
                     if (sigmaCornerOffset != 0.0) {
                         const float distance = sqrt(rtengine::SQR(i + tileSize / 2 - H / 2) + rtengine::SQR(j + tileSize / 2 - W / 2));
-                        const float sigmaTile = sigma + distanceFactor * distance;
+                        const float sigmaTile = static_cast<float>(sigma) + distanceFactor * distance;
                         if (sigmaTile >= 0.4f) {
                             float lkernel7[7][7];
-                            compute7x7kernel(sigma + distanceFactor * distance, lkernel7);
+                            compute7x7kernel(static_cast<float>(sigma) + distanceFactor * distance, lkernel7);
                             for (int k = 0; k < iterations - 1; ++k) {
                                 // apply 7x7 gaussian blur and divide luminance by result of gaussian blur
                                 gauss7x7div(tmpIThr, tmpThr, lumThr, fullTileSize, fullTileSize, lkernel7);
